Add kline support to Websockets API
Added a new function in the Websockets API for handling kline (or candlestick chart) data. It provides ways to parse these kline events and conduct related tests. Various time intervals for kline data are defined as constants. Besides, the timeout mechanism is introduced to the invoking of the API to avoid blocking.